HylaFAX, Version 3.0pl0 (release)
git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@3 5505949e-d877-4686-9e03-c53b7a51b376

+++ sgi2fax/imgtofax.c
@@ -1,7 +1,8 @@
-/*	$Header: /usr/people/sam/fax/sgi2fax/RCS/imgtofax.c,v 1.16 1994/05/16 19:27:59 sam Exp $ */
+/*	$Header: /usr/people/sam/fax/./sgi2fax/RCS/imgtofax.c,v 1.20 1995/04/08 21:43:32 sam Rel $ */
 /*
- * Copyright (c) 1990, 1991, 1992, 1993, 1994 Sam Leffler
- * Copyright (c) 1991, 1992, 1993, 1994 Silicon Graphics, Inc.
+ * Copyright (c) 1990-1995 Sam Leffler
+ * Copyright (c) 1991-1995 Silicon Graphics, Inc.
+ * HylaFAX is a trademark of Silicon Graphics
  *
  * Permission to use, copy, modify, distribute, and sell this software and 
  * its documentation for any purpose is hereby granted without fee, provided
@@ -116,7 +117,7 @@ tofax(short* wp, int n)
 	}
     }
     (void) TIFFWriteScanline(tif, row,
-	TIFFCurrentRow(tif) == -1 ? 0 : TIFFCurrentRow(tif), 0);
+	TIFFCurrentRow(tif) == (uint32) -1 ? 0 : TIFFCurrentRow(tif), 0);
 }
 
 static float transfunc(f)
